src/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.4
mysqlclient==2.2.7
pillow==10.3.0
pypng==0.20220715.0
qrcode==7.4.2
reportlab==4.2.0
sqlparse==0.5.0
typing_extensions==4.11.0
